+++ src/main/kotlin/racing/util/RandomNumberGenerator.kt
@@ -1,5 +1,13 @@
 package racing.util
 
-interface RandomNumberGenerator {
-    fun generate(): Int
+class RandomNumberGenerator : NumberGenerator {
+    override fun generate(): Int {
+        val range = MIN_RANGE..MAX_RANGE
+        return range.random()
+    }
+
+    companion object {
+        const val MIN_RANGE = 0
+        const val MAX_RANGE = 9
+    }
 }
